Move AIFunctionFactory down to M.E.AI.Abstractions
